Feature map (#68)
* search버튼-> 결과페이지 이동 && 가게리스트 불러온 뒤 무한스크롤방식구현(필터링X)

* Recommend -> 오늘의 식탁은?

* Object 변수를 이용한 url 파라미터 만들기

* 매장정보 파라미터 넘기기

* 파라미터 넘겨받은 값으로 매장상세페이지에서 지도 위치 띄우기

* 메인페이지에서 추천가게 누를 시 매장상세페이지로 이동 (파라미터넘겨받음)

* 검색결과페이지에서 사용자위치 나타나도록 하였음

* 매장상세페이지에서 매장위치핀과 사용자위치핀 같이 나오도록 함

* 마커 인포윈도우 표시하는 클릭이벤트

* 메인페이지에서 검색결과 파라미터로 넘김(필터링은 아직X)

* 검색결과 IF문처리

* Recommend -> 오늘의 식탁은?

* Object 변수를 이용한 url 파라미터 만들기

* 매장정보 파라미터 넘기기

* 파라미터 넘겨받은 값으로 매장상세페이지에서 지도 위치 띄우기

* 메인페이지에서 추천가게 누를 시 매장상세페이지로 이동 (파라미터넘겨받음)

* 검색결과페이지에서 사용자위치 나타나도록 하였음

* 매장상세페이지에서 매장위치핀과 사용자위치핀 같이 나오도록 함

* 마커 인포윈도우 표시하는 클릭이벤트

* 메인페이지에서 검색결과 파라미터로 넘김(필터링은 아직X)

* 검색결과 IF문처리

+++ src/main/java/com/todaysTable/vo/StoreVO.java
@@ -20,8 +20,8 @@ public class StoreVO {
 	private String break_time;
 	private String category;
 	private Date last_update_date;
-	private int latitude;
-	private int longitude;
+	private float latitude;
+	private float longitude;
 	private String admin_id;
 	
 }
